Moved interfaces to Domain
