+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Program.cs
@@ -4,6 +4,7 @@ using CleanArchitectureBlazor.Components.Account;
 using CleanArchitectureBlazor.Configuration;
 using CleanArchitectureBlazor.Data;
 using CleanArchitectureBlazor.Services;
+using Domain;
 using Microsoft.AspNetCore.Components.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
