+++ Domain/IStaffRepository.cs
@@ -1,6 +1,4 @@
-using Domain;
-
-namespace CleanArchitectureBlazor.Services;
+namespace Domain;
 
 /// <summary>
 /// Service for managing staff
@@ -9,10 +7,10 @@ public interface IStaffRepository
 {
     Task<List<Staff>> GetAllStaffAsync();
     Task<Staff?> GetStaffByIdAsync(int id);
-    Task<Staff> CreateStaffAsync(Domain.Staff staff);
-    Task<Staff> UpdateStaffAsync(Domain.Staff staff);
+    Task<Staff> CreateStaffAsync(Staff staff);
+    Task<Staff> UpdateStaffAsync(Staff staff);
     Task DeleteStaffAsync(int id);
-    Task<List<Domain.Staff>> GetActiveStaffAsync();
-    Task<List<Domain.Staff>> GetStaffByRoleAsync(StaffRole role);
+    Task<List<Staff>> GetActiveStaffAsync();
+    Task<List<Staff>> GetStaffByRoleAsync(StaffRole role);
     Task<bool> IsEmailUniqueAsync(string email, int? excludeId = null);
 }
